refactor: move MemPool to MemoryPool module
